+++ spec/virginia/service_spec.rb
@@ -18,7 +18,7 @@ describe Virginia::Service do
   end
 
   it "should instantiate the handler" do
-    rack_logger = mock 'Rack::CommonLogger'
+    rack_logger = double 'Rack::CommonLogger'
     ::Rack::CommonLogger.should_receive(:new).once.with(TestApp, Adhearsion.logger).and_return rack_logger
     ::Reel::Rack::Server.should_receive(:supervise_as).once.with(:reel_rack_server, rack_logger, options)
     Virginia::Service.start